search members and code improv
